+++ client/src/spFormat.ts
@@ -26,7 +26,7 @@ export class SMDocumentFormattingEditProvider
 		let UseTab : string = insert_spaces? "Never":"Always";
 		let tabSize : string = Workspace.getConfiguration("editor").get("tabSize");
 		
-		let default_styles : string[] = Workspace.getConfiguration("sourcepawnLanguageServer").get("formatterSettings");
+		let default_styles : string[] = Workspace.getConfiguration("sourcepawn").get("formatterSettings");
     
 		let default_style: string = "{" + default_styles.join(", ") + "}";